refactor(llmtools): split agent tools by responsibility
